+++ cw-orch-daemon/src/builder.rs
@@ -4,7 +4,7 @@ use crate::{
         base_sender::{SenderBuilder, SenderOptions},
         sender_trait::SenderTrait,
     },
-    DaemonAsync, DaemonBuilder, Wallet,
+    DaemonAsync, DaemonAsyncBase, DaemonBuilder, Wallet,
 };
 use std::sync::Arc;
 
@@ -12,7 +12,6 @@ use bitcoin::secp256k1::All;
 
 use super::{error::DaemonError, senders::base_sender::Sender, state::DaemonState};
 use cw_orch_core::environment::ChainInfoOwned;
-
 /// The default deployment id if none is provided
 pub const DEFAULT_DEPLOYMENT: &str = "default";
 
@@ -35,10 +34,13 @@ pub struct DaemonAsyncBuilderBase<SenderGen: SenderTrait> {
     // # Optional
     pub(crate) deployment_id: Option<String>,
 
-    /* Sender related options */
-    /// Wallet sender
-    /// Will be used in priority when set
-    pub(crate) sender: Option<SenderGen::SenderBuilder>,
+    pub(crate) sender: Option<SenderGen>,
+
+    // TODO, reallow rebuilding
+    // /* Sender related options */
+    // /// Wallet sender
+    // /// Will be used in priority when set
+    // pub(crate) sender: SenderGen::SenderBuilder,
     /// Specify Daemon Sender Options
     pub(crate) sender_options: SenderOptions,
 }
@@ -50,8 +52,8 @@ impl<SenderGen: SenderTrait> Default for DaemonAsyncBuilderBase<SenderGen> {
         Self {
             chain: Default::default(),
             deployment_id: Default::default(),
-            sender: Default::default(),
             sender_options: Default::default(),
+            sender: None,
         }
     }
 }
@@ -70,21 +72,22 @@ impl<SenderGen: SenderTrait> DaemonAsyncBuilderBase<SenderGen> {
         self
     }
 
-    /// Set the mnemonic to use with this chain.
-    /// Defaults to env variable depending on the environment.  
-    ///
-    /// Variables: LOCAL_MNEMONIC, TEST_MNEMONIC and MAIN_MNEMONIC
-    pub fn mnemonic(&mut self, mnemonic: impl ToString) -> &mut Self {
-        self.sender = Some(SenderBuilder::Mnemonic(mnemonic.to_string()));
-        self
-    }
+    // TODO
+    // /// Set the mnemonic to use with this chain.
+    // /// Defaults to env variable depending on the environment.
+    // ///
+    // /// Variables: LOCAL_MNEMONIC, TEST_MNEMONIC and MAIN_MNEMONIC
+    // pub fn mnemonic(&mut self, mnemonic: impl ToString) -> &mut Self {
+    //     self.sender = Some(SenderBuilder::Mnemonic(mnemonic.to_string()));
+    //     self
+    // }
 
-    /// Specifies a sender to use with this chain
-    /// This will be used in priority when set on the builder
-    pub fn sender(&mut self, wallet: Sender<All>) -> &mut Self {
-        self.sender = Some(SenderBuilder::Sender(wallet));
-        self
-    }
+    // /// Specifies a sender to use with this chain
+    // /// This will be used in priority when set on the builder
+    // pub fn sender(&mut self, wallet: Sender<All>) -> &mut Self {
+    //     self.sender = Some(SenderBuilder::Sender(wallet));
+    //     self
+    // }
 
     /// Specifies whether authz should be used with this daemon
     pub fn authz_granter(&mut self, granter: impl ToString) -> &mut Self {
@@ -105,7 +108,7 @@ impl<SenderGen: SenderTrait> DaemonAsyncBuilderBase<SenderGen> {
     }
 
     /// Build a daemon
-    pub async fn build(&self) -> Result<DaemonAsync, DaemonError> {
+    pub async fn build(&self) -> Result<DaemonAsyncBase<SenderGen>, DaemonError> {
         let chain = self
             .chain
             .clone()
@@ -114,26 +117,17 @@ impl<SenderGen: SenderTrait> DaemonAsyncBuilderBase<SenderGen> {
             .deployment_id
             .clone()
             .unwrap_or(DEFAULT_DEPLOYMENT.to_string());
+
         let state = Arc::new(DaemonState::new(chain, deployment_id, false).await?);
         // if mnemonic provided, use it. Else use env variables to retrieve mnemonic
-        let sender_options = self.sender_options.clone();
 
-        let sender = match self.sender.clone() {
-            Some(sender) => match sender {
-                SenderBuilder::Mnemonic(mnemonic) => {
-                    Sender::from_mnemonic_with_options(&state, &mnemonic, sender_options)?
-                }
-                SenderBuilder::Sender(mut sender) => {
-                    sender.set_options(self.sender_options.clone());
-                    sender
-                }
-            },
-            None => Sender::new_with_options(&state, sender_options)?,
+        let sender = if let Some(sender) = self.sender.clone() {
+            sender
+        } else {
+            SenderGen::build(self.sender_options.clone(), &state).map_err(Into::into)?
         };
-        let daemon = DaemonAsync {
-            state,
-            sender: Arc::new(sender),
-        };
+
+        let daemon = DaemonAsyncBase { state, sender };
         print_if_log_disabled()?;
         Ok(daemon)
     }