+++ cw-orch-daemon/src/senders/mod.rs
@@ -1,2 +1,3 @@
 pub mod base_sender;
+pub mod multiple_sender;
 pub mod sender_trait;