+++ cw-orch-daemon/tests/common/mod.rs
@@ -17,6 +17,7 @@ mod node {
 
     // Defaults for env vars
     const CONTAINER_NAME: &str = "juno_node_1";
+    // From https://github.com/CosmosContracts/juno/blob/32568dba828ff7783aea8cb5bb4b8b5832888255/docker/test-user.env#L2
     const LOCAL_MNEMONIC: &str = "clip hire initial neck maid actor venue client foam budget lock catalog sweet steak waste crater broccoli pipe steak sister coyote moment obvious choose";
 
     use uid::Id as IdT;