+++ packages/cw-orch-osmosis-test-tube/src/queriers/node.rs
@@ -1,14 +1,13 @@
 use std::{cell::RefCell, rc::Rc};
 
-use crate::mock::cw_multi_test::AppResponse;
+use crate::OsmosisTestTube;
 use cosmwasm_std::{BlockInfo, Timestamp};
 use cw_orch_core::{
     environment::{NodeQuerier, Querier, QuerierGetter, StateInterface},
     CwEnvError,
 };
+use cw_orch_mock::cw_multi_test::AppResponse;
 use osmosis_test_tube::OsmosisTestApp;
-
-use crate::prelude::OsmosisTestTube;
 
 pub struct OsmosisTestTubeNodeQuerier {
     app: Rc<RefCell<OsmosisTestApp>>,