+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## Unreleased
 
+- Added a test to make sure the derive macros stay compatible with new cw-orch versions
+- Changed the derive macros import from cw_orch to cw_orch_core. This allows changing the cw-orch API without breaking the derive macros.
+- Cw-orch mock env info doesn't error when using chain ids that don't match the `osmosis-1` pattern
+
+### Breaking
+
+- Daemon : Changed return types on daemon queriers to match Cosmwasm std types
+- Cw-orch : Separate osmosis test tube from cw-orch. Its not available in its own crate `cw-orch-osmosis-test-tube`
+
+## 0.22.0
+
 - Updated osmosis test tube to 24.0.1 ,that avoids re-compiling osmosis test tube
 - Added `balance` query at the root of QueryHandler
 - Added DaemonBuilder configuration for grpc url and fee overwriting
@@ -14,6 +25,7 @@
 ## 0.21.2
 
 - Allow cw-orch wasm compilation without features
+- Transaction Response now inspects logs and events to find matching events.
 
 ## 0.21.1